At least, – for the purposes of this library – I don't see why there'd be any need to grab the header info. for any returned calls: we really just want the JSON. Plus, I'd think, the header info., as plaintext, would be harder to parse reliably. I can always change if I find I'm mistaken, down the road.
I'm going to make this synchronous since, if I'm trying to register
the app. – to begin with –, I'm likely not going to want the user
wandering off and changing buffers before I suddenly need to switch
and setup the app.'s buffer but, also, because it shouldn't be that
long of a wait. But it also shouldn't be surprising you're going to
have to wait for the app. to "load".